+++ blotztask-mobile/index.ts
@@ -3,9 +3,9 @@ import { Platform } from "react-native";
 
 if (Platform.OS === "android") {
   type AndroidWidgetRegistration =
-    typeof import("./src/feature/widget/android/register-android-widget-task-handler");
+    typeof import("./src/feature/widget/android/services/register-android-widget-task-handler");
   const { registerAndroidWidgetTaskHandler } =
-    require("./src/feature/widget/android/register-android-widget-task-handler") as AndroidWidgetRegistration;
+    require("./src/feature/widget/android/services/register-android-widget-task-handler.ts") as AndroidWidgetRegistration;
 
   registerAndroidWidgetTaskHandler();
 }
