+++ blotztask-mobile/src/feature/widget/android/services/register-android-widget-task-handler.ts
@@ -1,5 +1,5 @@
 import { registerWidgetTaskHandler } from "react-native-android-widget";
-import { widgetTaskHandler } from "@/feature/widget/android/widget-task-handler";
+import { widgetTaskHandler } from "@/feature/widget/android/handlers/widget-task-handler";
 
 export function registerAndroidWidgetTaskHandler(): void {
   registerWidgetTaskHandler(widgetTaskHandler);
