+++ blotztask-mobile/src/feature/widget/android/services/today-tasks-widget-cache-storage.ts
@@ -1,4 +1,5 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
+
 import type { TaskWidgetCache } from "@/feature/widget/models/task-widget-cache";
 import type {
   TaskWidgetSnapshotItem,
@@ -6,7 +7,7 @@ import type {
   TasksWidgetSnapshot,
 } from "@/feature/widget/models/today-tasks-widget-snapshot";
 
-const TODAY_TASKS_WIDGET_CACHE_KEY = "blotztask.widget.todayTasksCache.v2";
+const TODAY_TASKS_WIDGET_CACHE_KEY = "blotztask.widget.todayTasksCache.v1";
 
 export async function writeTodayTasksWidgetCache(cache: TaskWidgetCache): Promise<void> {
   await AsyncStorage.setItem(TODAY_TASKS_WIDGET_CACHE_KEY, JSON.stringify(cache));
@@ -20,7 +21,7 @@ export async function readTodayTasksWidgetCache(): Promise<TaskWidgetCache | nul
     const parsed: unknown = JSON.parse(rawCache);
     return isTaskWidgetCache(parsed) ? parsed : null;
   } catch {
-    return null;
+    return null; // may need to throw error
   }
 }
 
@@ -29,10 +30,10 @@ function isTaskWidgetCache(value: unknown): value is TaskWidgetCache {
     return false;
   }
 
-  return Object.values(value.days).every(isTodayTasksWidgetSnapshot);
+  return Object.values(value.days).every(isTasksWidgetSnapshot);
 }
 
-function isTodayTasksWidgetSnapshot(value: unknown): value is TasksWidgetSnapshot {
+function isTasksWidgetSnapshot(value: unknown): value is TasksWidgetSnapshot {
   if (!isRecord(value)) return false;
 
   return (
