+++ blotztask-mobile/src/feature/widget/services/register-today-tasks-widget-layout.ios.ts
@@ -1,5 +1,5 @@
 export function registerTodayTasksWidgetLayout(): void {
-  void import("@/feature/widget/ios/today-tasks-widget").catch((error: unknown) => {
+  void import("@/feature/widget/ios/components/today-tasks-widget").catch((error: unknown) => {
     if (__DEV__) {
       console.warn("[TodayTasksWidget] Failed to register widget layout", error);
     }
