+++ blotztask-mobile/src/feature/widget/services/today-tasks-widget-sync.ts
@@ -1,12 +1,19 @@
 import { Platform } from "react-native";
 import type { TaskWidgetCache } from "@/feature/widget/models/task-widget-cache";
 
-export function syncTodayTasksWidgetCache(cache: TaskWidgetCache): void {
-  if (Platform.OS !== "android") return;
+export async function syncTodayTasksWidgetCache(cache: TaskWidgetCache): Promise<void> {
+  if (Platform.OS === "ios") {
+    const { syncIosTodayTasksWidgetCache } =
+      await import("@/feature/widget/ios/services/sync-ios-today-tasks-widget-cache");
 
-  void import("@/feature/widget/services/today-tasks-widget-sync-android").then(
-    ({ syncTodayTasksWidgetCache: syncAndroidCache }) => {
-      syncAndroidCache(cache);
-    },
-  );
+    await syncIosTodayTasksWidgetCache(cache);
+    return;
+  }
+
+  if (Platform.OS === "android") {
+    const { syncAndroidTodayTasksWidgetCache } =
+      await import("@/feature/widget/android/services/sync-android-today-tasks-widget-cache");
+
+    await syncAndroidTodayTasksWidgetCache(cache);
+  }
 }
